requirements.txt
setuptools~=68.2.2
pymysql~=1.1.0
python~=3.8.18
os
python-dotenv
